git subrepo commit (merge) kigs
subrepo:
  subdir:   "kigs"
  merged:   "225c82cd89"
upstream:
  origin:   "https://github.com/assoria/kigs.git"
  branch:   "master"
  commit:   "e8b68d77af"
git-subrepo:
  version:  "0.4.1"
  origin:   "???"
  commit:   "???"

Former-commit-id: a5e6c54b825bb654b4eb9a426c371e2c57669b02
Former-commit-id: cfcc335e00279833f5c1a25a87faf305f11e49f8

+++ framework/2DLayers/Sources/BaseUI2DLayer.cpp
@@ -324,7 +324,7 @@ void BaseUI2DLayer::TravDraw(TravState* state)
 
 				if (compare_stencil_stacks())
 				{
-					renderer->SetStencilMode(RendererCullMode::RENDERER_CULL_FRONT_AND_BACK, RENDERER_STENCIL_EQUAL, cache_stencil_stack.size(), 0xFFFFFFFF);
+					renderer->SetStencilMode(RendererCullMode::RENDERER_CULL_FRONT_AND_BACK, RENDERER_STENCIL_EQUAL, (int)cache_stencil_stack.size(), 0xFFFFFFFF);
 				}
 				else
 				{
