+++ framework/2DLayers/Sources/UI/UINode3DLayer.cpp
@@ -336,7 +336,7 @@ void UINode3DLayer::TravDraw(TravState* state)
 
 				if (compare_stencil_stacks())
 				{
-					renderer->SetStencilMode(RendererCullMode::RENDERER_CULL_FRONT_AND_BACK, RENDERER_STENCIL_EQUAL, cache_stencil_stack.size(), 0xFFFFFFFF);
+					renderer->SetStencilMode(RendererCullMode::RENDERER_CULL_FRONT_AND_BACK, RENDERER_STENCIL_EQUAL, (int)cache_stencil_stack.size(), 0xFFFFFFFF);
 				}
 				else
 				{
